% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsmethods-package.R
\docType{package}
\name{phsmethods-package}
\alias{phsmethods}
\alias{phsmethods-package}
\title{phsmethods: Standard Methods for Use in Public Health Scotland}
\description{
A collection of methods for commonly undertaken analytical tasks, primarily developed for Public Health Scotland (PHS) analysts, but the package is also generally useful to others working in the healthcare space, particularly since it has functions for working with Community Health Index (CHI) numbers. The package can help to make data manipulation and analysis more efficient and reproducible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Public-Health-Scotland/phsmethods}
  \item \url{https://code.publichealthscotland.scot/phsmethods}
  \item Report bugs at \url{https://github.com/Public-Health-Scotland/phsmethods/issues}
}

}
\author{
\strong{Maintainer}: Tina Fu \email{Yuyan.Fu2@phs.scot}

Authors:
\itemize{
  \item Tina Fu \email{Yuyan.Fu2@phs.scot}
  \item David Caldwell \email{David.Caldwell@phs.scot}
  \item Jack Hannah \email{jack.hannah2@phs.scot}
  \item Ciara Gribben \email{Ciara.Gribben@phs.scot}
  \item Chris Deans \email{Chris.Deans2@phs.scot}
  \item Jaime Villacampa \email{Jaime.Villacampa@phs.scot}
  \item Graeme Gowans \email{Graeme.Gowans@phs.scot}
  \item James Hayes \email{James.Hayes2@phs.scot} (\href{https://orcid.org/0000-0002-5380-2029}{ORCID})
  \item Nicolaos Christofidis \email{nicolaos.christofidis@phs.scot}
}

Other contributors:
\itemize{
  \item Public Health Scotland \email{phs.datascience@phs.scot} [copyright holder]
  \item Lucinda Lawrie \email{Lucinda.Lawrie@phs.scot} [reviewer]
  \item Alice Byers [contributor]
  \item Alan Yeung \email{Alan.Yeung@phs.scot} [contributor]
}

}
\keyword{internal}
